+++ peer/roundrobin/peerring.go
@@ -26,6 +26,7 @@ import (
 
 	"go.uber.org/yarpc/api/peer"
 	"go.uber.org/yarpc/api/transport"
+	"go.uber.org/yarpc/peer/peerlist/v2"
 	"go.uber.org/yarpc/yarpcerrors"
 )
 
@@ -51,9 +52,11 @@ type peerRing struct {
 	nextNode *ring.Ring
 }
 
+var _ peerlist.Implementation = (*peerRing)(nil)
+
 // Add a peer.StatusPeer to the end of the peerRing, if the ring is empty it
 // initializes the nextNode marker
-func (pr *peerRing) Add(p peer.StatusPeer) peer.Subscriber {
+func (pr *peerRing) Add(p peer.StatusPeer, _ peer.Identifier) peer.Subscriber {
 	sub := &subscriber{peer: p}
 	newNode := ring.New(1)
 	newNode.Value = sub
@@ -71,7 +74,7 @@ func (pr *peerRing) Add(p peer.StatusPeer) peer.Subscriber {
 
 // Remove the peer from the ring. Use the subscriber to address the node of the
 // ring directly.
-func (pr *peerRing) Remove(p peer.StatusPeer, s peer.Subscriber) {
+func (pr *peerRing) Remove(p peer.StatusPeer, _ peer.Identifier, s peer.Subscriber) {
 	sub, ok := s.(*subscriber)
 	if !ok {
 		// Don't panic.