+++ transport/http/transport.go
@@ -21,6 +21,7 @@
 package http
 
 import (
+	"math/rand"
 	"net"
 	"net/http"
 	"sync"
@@ -45,6 +46,8 @@ type transportOptions struct {
 	responseHeaderTimeout time.Duration
 	connTimeout           time.Duration
 	connBackoffStrategy   backoffapi.Strategy
+	innocenceWindow       time.Duration
+	jitter                func(int64) int64
 	tracer                opentracing.Tracer
 	buildClient           func(*transportOptions) *http.Client
 	logger                *zap.Logger
@@ -56,6 +59,8 @@ var defaultTransportOptions = transportOptions{
 	connTimeout:         defaultConnTimeout,
 	connBackoffStrategy: backoff.DefaultExponential,
 	buildClient:         buildHTTPClient,
+	innocenceWindow:     defaultInnocenceWindow,
+	jitter:              rand.Int63n,
 }
 
 func newTransportOptions() transportOptions {
@@ -160,6 +165,24 @@ func ConnBackoff(s backoffapi.Strategy) TransportOption {
 	}
 }
 
+// InnocenceWindow is the duration after the peer connection management loop
+// will suspend suspicion for a peer after successfully checking whether the
+// peer is live with a fresh TCP connection.
+//
+// The default innocence window is 5 seconds.
+//
+// A timeout does not necessarily indicate that a peer is unavailable,
+// but it could indicate that the connection is half-open, that the peer died
+// without sending a TCP FIN packet.
+// In this case, the peer connection management loop attempts to open a TCP
+// connection in the background, once per innocence window, while suspicious of
+// the connection, leaving the peer available until it fails.
+func InnocenceWindow(d time.Duration) TransportOption {
+	return func(options *transportOptions) {
+		options.innocenceWindow = d
+	}
+}
+
 // Tracer configures a tracer for the transport and all its inbounds and
 // outbounds.
 func Tracer(tracer opentracing.Tracer) TransportOption {
@@ -204,6 +227,8 @@ func (o *transportOptions) newTransport() *Transport {
 		client:              o.buildClient(o),
 		connTimeout:         o.connTimeout,
 		connBackoffStrategy: o.connBackoffStrategy,
+		innocenceWindow:     o.innocenceWindow,
+		jitter:              o.jitter,
 		peers:               make(map[string]*httpPeer),
 		tracer:              o.tracer,
 		logger:              logger,
@@ -244,6 +269,8 @@ type Transport struct {
 	connTimeout         time.Duration
 	connBackoffStrategy backoffapi.Strategy
 	connectorsGroup     sync.WaitGroup
+	innocenceWindow     time.Duration
+	jitter              func(int64) int64
 
 	tracer opentracing.Tracer
 	logger *zap.Logger