+++ yarpctest/fake_outbound.go
@@ -113,15 +113,15 @@ func (o *FakeOutbound) Call(ctx context.Context, req *transport.Request) (*trans
 	if o.callOverride != nil {
 		return o.callOverride(ctx, req)
 	}
-	return nil, fmt.Errorf(`no outbound callable specified on the outbound`)
+	return nil, fmt.Errorf(`no outbound callable specified on the fake outbound`)
 }
 
 // CallOneway pretends to send a oneway RPC, but actually just returns an error.
 func (o *FakeOutbound) CallOneway(ctx context.Context, req *transport.Request) (transport.Ack, error) {
-	return nil, fmt.Errorf(`call oneway outbound is unsupported`)
+	return nil, fmt.Errorf(`fake outbound does not support call oneway`)
 }
 
 // CallStream pretends to send a Stream RPC, but actually just returns an error.
 func (o *FakeOutbound) CallStream(ctx context.Context, req *transport.StreamRequest) (*transport.ClientStream, error) {
-	return nil, fmt.Errorf(`call stream outbound is unsupported`)
+	return nil, fmt.Errorf(`fake outbound does not support call stream`)
 }